🔨 add script for filling user data
