+++ package.json
@@ -12,7 +12,8 @@
     "db:generate": "set -a; . ./.env.local; set +a && drizzle-kit generate:pg",
     "db:studio": "set -a; . ./.env.local; set +a && drizzle-kit studio --port 3003 --verbose",
     "db:migrate": "set -a; . ./.env.local; set +a && tsx ./migrate.ts",
-    "format": "prettier . --write"
+    "format": "prettier . --write",
+    "db:fetch-users": "set -a; . ./.env.local; set +a && tsx ./scripts/fetchUsers.ts"
   },
   "type": "module",
   "dependencies": {
