+++ src/app/(actions)/theme.ts
@@ -1,7 +1,6 @@
 "use server";
 
 import { z } from "zod";
-import { withAuth } from "~/lib/server/utils.server";
 import { getSession } from "./auth";
 import { cache } from "react";
 import { updateUserTheme } from "~/app/(models)/user";
@@ -10,6 +9,7 @@ import { revalidatePath } from "next/cache";
 import { users } from "~/lib/server/db/schema/user.sql";
 import { createSelectSchema } from "drizzle-zod";
 import { redirect } from "next/navigation";
+import { withAuth } from "~/lib/server/rsc-utils.server";
 
 const userThemeSchema = createSelectSchema(users).pick({
   preferred_theme: true
